PIYUSH | modal open logic change
